+++ src/TSLite.Server/Json/ServerJsonContext.cs
@@ -25,6 +25,7 @@ namespace TSLite.Server.Json;
 [JsonSerializable(typeof(HealthResponse))]
 [JsonSerializable(typeof(LoginRequest))]
 [JsonSerializable(typeof(LoginResponse))]
+[JsonSerializable(typeof(BulkIngestResponse))]
 [JsonSerializable(typeof(MetricsSnapshotEvent))]
 [JsonSerializable(typeof(SlowQueryEvent))]
 [JsonSerializable(typeof(DatabaseEvent))]